refactor(telegram): move crash-recovery resume to resume.rs, verbatim (#471 phase 1)
Sixth mechanical cut: resume_session (647 lines) moves to resume.rs.
Pure move; visibility widened on the handler internals it uses
(TypingGuard, tool_context, drain_remaining_display,
StreamingState::render), re-export keeps the ui.rs call path stable.

Phase 1 tally: handler.rs 7221 → 4547 lines; new modules flow (928),
markdown (367), media (319), intermediates (257), keyboards (243),
resume (656) — all under the 1k ceiling, most under 400. The remaining
4547 is dominated by the single handle_message function, which shrinks
further in phase 2 (delivery extraction, #471) and a follow-up front-
half split.
